Refactor ModelBuilders for COTS and PIV
The wrapper for error/logger/newrelic management is refactored:
* Removed from abstract_scnf_resource
* Placed into abstract_builder.wrap_build
* Used by PIV and COTS
* Prepared to be shared also by GTFS-RT (TODO)

AbstractKirinModelBuilder is created to explicit what is expected by
wrap_build when using a KirinModelBuilder.
It also automatically manages the navitia_wrapper.

abstract_sncf_model_maker is removed as cots.KirinModelMaker now
inherits from AbstractKirinModelBuilder:
* move functions into COTS builder file
* common functions with PIV might be shared later (but no inheritance)
