+++ kirin/piv/__init__.py
@@ -30,3 +30,4 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
+from kirin.piv.model_maker import KirinModelBuilder
